@@ -1,38 +1,103 @@
-[tool.poetry]
+[project]
 name = "avgangstider"
-version = "0.1.3"
+version = "0.9.0"
 description = "Avgangstider fra EnTur"
-authors = ["Martin Høy <marhoy@gmail.com>"]
+readme = "README.rst"
+authors = [{ name = "Martin Høy", email = "martin@hoy.priv.no" }]
+requires-python = ">=3.10"
+dependencies = [
+    "flask>=3.1.0",
+    "gunicorn>=23.0.0",
+    "loguru>=0.7.3",
+    "requests>=2.32.3",
+]
 
-[tool.poetry.dependencies]
-python = "^3.7"
-flask = "^1.1"
-requests = "^2.22"
-gunicorn = { version = "^20.0", optional = true }
-
-[tool.poetry.dev-dependencies]
-pytest = "*"
-pytest-cov = "*"
-pytest-flask = "*"
-black = "*"
-flake8 = "*"
-isort = "*"
-mypy = "*"
-types-requests = "*"
-tox = "*"
-sphinx = "*"
-sphinx-rtd-theme = "*"
-
-[tool.poetry.extras]
-gunicorn = ["gunicorn"]
+[dependency-groups]
+docs = ["sphinx>=8.1.3", "sphinx-rtd-theme>=3.0.2"]
+dev = [
+    "ipykernel>=6.29.5",
+    "pre-commit>=4.1.0",
+    "pytest-cov>=6.0.0",
+    "pytest-flask>=1.3.0",
+]
 
 [build-system]
-requires = ["poetry>=1.0"]
-build-backend = "poetry.masonry.api"
+requires = ["hatchling"]
+build-backend = "hatchling.build"
 
-[tool.isort]
-profile = "black"
 
-[[tool.mypy.overrides]]
-module = ["flask", "werkzeug.datastructures"]
-ignore_missing_imports = true
+[tool.ruff]
+# Consider packages in the src directory as first-party code
+src = ["src"]
+# Run ruff on Notebooks as well
+extend-include = ["*.ipynb"]
+
+[tool.ruff.lint.per-file-ignores]
+# Allow the use of assert in tests
+"tests/*" = ["S101"]
+
+[tool.ruff.lint]
+select = [
+    "F",   # Pyflakes
+    "E",   # pycodestyle
+    "W",   # pycodestyle
+    "C90", # mccabe
+    "I",   # isort
+    "ICN", # flake8-import-conventions
+    "N",   # pep8-naming
+    "D",   # pydocstyle
+    "UP",  # pyupgrade
+    "ANN", # flake8-annotations
+    "S",   # flake8-bandit
+    "B",   # flake8-bugbear
+    "A",   # flake8-builtins
+    "C4",  # flake8-comprehensions
+    "DTZ", # flake8-datetimez
+    "FA",  # flake8-future-annotations
+    "ISC", # flake8-implicit-str-concat
+    "ICN", # flake8-import-conventions
+    "PT",  # flake8-pytest-style
+    "Q",   # flake8-quotes
+    "RET", # flake8-return
+    "SIM", # flake8-simplify
+    "TID", # flake8-tidy-imports
+    "ARG", # flake8-unused-arguments
+    "PTH", # flake8-use-pathlib
+    "PD",  # pandas-vet
+    "PLC", # Pylint convention
+    "PLE", # Pylint error
+    # "PLR",  # Pylint refactor
+    "NPY",  # NumPy
+    "FAST", # FastAPI
+    "RUF",  # Ruff-specific rules
+]
+ignore = []
+
+[tool.ruff.lint.pydocstyle]
+convention = "google"
+
+
+# mypy configuration, taken from pydantic documentation
+[tool.mypy]
+plugins = ["pydantic.mypy"]
+
+follow_imports = "silent"
+warn_redundant_casts = true
+warn_unused_ignores = true
+disallow_any_generics = true
+check_untyped_defs = true
+no_implicit_reexport = true
+
+disallow_untyped_defs = true
+
+[tool.pydantic-mypy]
+init_forbid_extra = true
+init_typed = true
+warn_required_dynamic_aliases = true
+
+
+# pytest configuration
+[tool.pytest.ini_options]
+addopts = "--cov-report term-missing:skip-covered --cov=avgangstider"
+# Ignore deprecation warnings from third party packages
+filterwarnings = []
