@@ -1,3 +1,7 @@
+"""Classes for the avgangstider package."""
+
+from __future__ import annotations
+
 import functools
 from dataclasses import dataclass
 from datetime import datetime
@@ -5,7 +9,7 @@ from datetime import datetime
 
 @dataclass
 class Departure:
-    """A data class to hold information about a departure"""
+    """A data class to hold information about a departure."""
 
     line_id: str
     line_name: str
@@ -17,30 +21,30 @@ class Departure:
 
     @property
     def departure_string(self) -> str:
-        """The departure time as a string relative to now()"""
+        """The departure time as a string relative to now()."""
         # How long is it to the departure?
         now = datetime.now(tz=self.departure_datetime.tzinfo)
         minutes = (self.departure_datetime - now).total_seconds() // 60
         if minutes <= 0:
             departure_string = "nå"
         elif minutes <= 30:
-            departure_string = "{:.0f} min".format(minutes)
+            departure_string = f"{minutes:.0f} min"
         else:
             departure_string = self.departure_datetime.strftime("%H:%M")
 
         return departure_string
 
-    # What a departure looks like if ypu print() it
-    def __str__(self):
-        return "{:2s} -> {:15s} @ {}".format(
-            self.line_name, self.destination, self.departure_string
+    def __str__(self) -> str:
+        """What a departure looks like if you print() it."""
+        return (
+            f"{self.line_name:2s} -> {self.destination:15s} @ {self.departure_string}"
         )
 
 
 @dataclass
 @functools.total_ordering
 class Situation:
-    """A data class to hold information about a situation"""
+    """A data class to hold information about a situation."""
 
     line_id: str
     line_name: str
@@ -49,14 +53,16 @@ class Situation:
     fg_color: str
     summary: str
 
-    # Define what it takes for two Situations to be equal
-    def __eq__(self, other):
+    def __eq__(self, other: object) -> bool:
+        """Define what it takes for two Situations to be equal."""
+        if not isinstance(other, Situation):
+            return NotImplemented
         return (self.line_name, self.summary) == (other.line_name, other.summary)
 
-    # Define what it takes for one Situation to be less than another
-    def __lt__(self, other):
+    def __lt__(self, other: Situation) -> bool:
+        """Define what it takes for one Situation to be less than another."""
         return (self.line_name, self.summary) < (other.line_name, other.summary)
 
-    # What a situation looks like if ypu print() it
-    def __str__(self):
-        return "{}: {}".format(self.line_name, self.summary)
+    def __str__(self) -> str:
+        """What a situation looks like if you print() it."""
+        return f"{self.line_name}: {self.summary}"
