@@ -1,9 +1,11 @@
+"""Utility functions for the avgangstider package."""
+
 import re
 from datetime import datetime
 
 
 def iso_str_to_datetime(timestamp_str: str) -> datetime:
-    """Convert timestamp string to datetime object
+    """Convert timestamp string to datetime object.
 
     Args:
         timestamp_str (str): A timestamp in iso format
