@@ -1,19 +1,26 @@
+"""Tests for the Flask app."""
+
+from collections.abc import Callable
 from unittest import mock
 
+from flask.testing import FlaskClient
 
-def test_static(client):
+from avgangstider.classes import Situation
+
+
+def test_static(client: FlaskClient) -> None:  # noqa: D103
     response = client.get("/static/css/dinfont.css")
     assert response.status_code == 200
     assert b"font-face" in response.data
 
 
-def test_app_without_stop_id(client):
+def test_app_without_stop_id(client: FlaskClient) -> None:  # noqa: D103
     response = client.get("/", follow_redirects=True)
     assert response.status_code == 200
     assert b"Examples" in response.data
 
 
-def test_app(client):
+def test_app(client: FlaskClient) -> None:  # noqa: D103
     response = client.get("/", query_string={"stop_id": "NSR:StopPlace:58366"})
     assert response.status_code == 200
     assert b"DepartureTable" in response.data
@@ -24,7 +31,7 @@ def test_app(client):
     assert b"DepartureTable" in response.data
 
 
-def test_departure_table(client):
+def test_departure_table(client: FlaskClient) -> None:  # noqa: D103
     # Get departures with an empty stop_id
     response = client.get("/departure_table", query_string={"stop_id": None})
     assert response.status_code == 200
@@ -43,7 +50,7 @@ def test_departure_table(client):
     assert response.status_code == 200
 
 
-def test_deviations(client):
+def test_deviations(client: FlaskClient) -> None:  # noqa: D103
     # Try first without having got any departures
     # There should be no situations
     response = client.get(
@@ -61,7 +68,7 @@ def test_deviations(client):
     assert response.status_code == 200
 
 
-def test_deviations_with_line(client):
+def test_deviations_with_line(client: FlaskClient) -> None:  # noqa: D103
     response = client.get(
         "/deviations",
         query_string={"stop_id": "NSR:StopPlace:58366", "line_id": "RUT:Line:3"},
@@ -70,9 +77,13 @@ def test_deviations_with_line(client):
 
 
 @mock.patch("avgangstider.get_situations")
-def test_deviations_mocked(mocked_func, client, saved_situations_list):
+def test_deviations_mocked(  # noqa: D103
+    mocked_func: Callable[[list[str], str], list[Situation]],
+    client: FlaskClient,
+    saved_situations_list: list[Situation],
+) -> None:
     # Load some mocked situations
-    mocked_func.return_value = saved_situations_list
+    mocked_func.return_value = saved_situations_list  # type: ignore
 
     client.get("/departure_table", query_string={"stop_id": "NSR:StopPlace:58366"})
 
